src/requirements.txt
astroid==2.3.3
colorama==0.4.3
cycler==0.10.0
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.2
mccabe==0.6.1
np==1.0.2
numpy==1.18.0
opencv-python==4.1.2.30
Pillow==6.2.1
pylint==2.4.4
pyparsing==2.4.6
pytesseract==0.3.1
python-dateutil==2.8.1
scipy==1.4.1
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
